refactor(liferay-cli): move sample CSS generation to a common facet
